Use Aqua BaseParser
